# backend-go/docs/swagger.yaml
definitions:
  github_com_lingmirror_backend-go_internal_response.PageResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  github_com_lingmirror_backend-go_internal_response.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_domain_sku.Product:
    properties:
      ai_description:
        type: string
      ai_status:
        type: string
      ai_title:
        type: string
      brand_id:
        type: integer
      cargo_type:
        type: string
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          type: integer
        type: array
      main_image:
        type: string
      name:
        type: string
      package_height_cm:
        type: number
      package_length_cm:
        type: number
      package_weight_kg:
        type: number
      package_width_cm:
        type: number
      platform_statuses:
        items:
          type: integer
        type: array
      product_height_cm:
        type: number
      product_length_cm:
        type: number
      product_weight_kg:
        type: number
      product_width_cm:
        type: number
      seo_keywords:
        items:
          type: integer
        type: array
      status:
        type: integer
      subtitle:
        type: string
      unit:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Get paginated list of products
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Filter by category
        in: query
        name: category_id
        type: integer
      - description: Filter by brand
        in: query
        name: brand_id
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.PageResult'
      security:
      - BearerAuth: []
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Product data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_sku.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - products
  /products/{id}:
    delete:
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - products
    get:
      description: Get a single product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Get product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_sku.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - products
  /skus:
    get:
      consumes:
      - application/json
      description: Get paginated list of SKUs
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Filter by product ID
        in: query
        name: product_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.PageResult'
      security:
      - BearerAuth: []
      summary: List SKUs
      tags:
      - products
swagger: "2.0"